ingestion_pipeline/requirements.txt
neo4j
weaviate-client
sentence-transformers
spacy
faker
tqdm